requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=0.24.2
matplotlib>=3.4.2
seaborn>=0.11.1
jupyter>=1.0.0
pyyaml>=5.4.1
xgboost>=1.4.2
lightgbm>=3.2.1
catboost>=0.26
shap>=0.40.0
optuna>=2.10.0 